fix fees to follow spec for execute fulfillment using unfinished estimated cost of execution
