+++ pallets/randomness/src/lib.rs
@@ -22,6 +22,11 @@ use frame_support::pallet;
 
 pub use pallet::*;
 
+pub mod traits;
+pub mod types;
+pub use traits::*;
+pub use types::*;
+
 // pub mod weights;
 // use weights::WeightInfo;
 // #[cfg(any(test, feature = "runtime-benchmarks"))]
@@ -33,163 +38,13 @@ pub use pallet::*;
 
 #[pallet]
 pub mod pallet {
+	use super::*;
 	// use crate::WeightInfo;
 	use frame_support::pallet_prelude::*;
-	use frame_support::traits::{Currency, ExistenceRequirement::KeepAlive, ReservableCurrency};
+	use frame_support::traits::{Currency, ReservableCurrency};
+	use frame_support::weights::WeightToFeePolynomial;
 	use frame_system::pallet_prelude::*;
-	use sp_runtime::traits::{CheckedSub, Saturating};
-
-	#[derive(PartialEq, Copy, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
-	/// Randomness storage item from BABE
-	pub enum BabeRandomness {
-		OneEpochAgo,
-		TwoEpochsAgo,
-		CurrentBlock,
-	}
-
-	#[derive(PartialEq, Copy, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
-	/// Information regarding the request type
-	pub enum RequestType<BlockNumber> {
-		/// BABE randomness from reading relay chain state proof
-		Babe {
-			info: BabeRandomness,
-			when: BlockNumber,
-		},
-		/// Per block VRF randomness
-		Local { when: BlockNumber },
-	}
-
-	impl<BlockNumber: Copy> RequestType<BlockNumber> {
-		pub fn when(&self) -> BlockNumber {
-			match self {
-				RequestType::Babe { when, .. } => *when,
-				RequestType::Local { when } => *when,
-			}
-		}
-	}
-
-	#[derive(PartialEq, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
-	#[scale_info(skip_type_params(T))]
-	pub struct Request<T: Config> {
-		/// Fee is returned to this account upon execution
-		pub refund_address: T::AccountId,
-		/// Contract that consumes the randomness
-		pub contract_address: T::AccountId,
-		/// Fee to pay for execution
-		pub fee: BalanceOf<T>,
-		/// Salt to use once randomness is ready
-		pub salt: T::Hash,
-		/// Details regarding request type and when it is due
-		pub info: RequestType<T::BlockNumber>,
-	}
-
-	impl<T: Config> Request<T> {
-		fn can_be_fulfilled(&self) -> bool {
-			self.info.when() <= frame_system::Pallet::<T>::block_number()
-		}
-	}
-
-	#[derive(PartialEq, Clone, Encode, Decode, RuntimeDebug, TypeInfo)]
-	#[scale_info(skip_type_params(T))]
-	pub struct RequestState<T: Config> {
-		/// Fee is returned to this account upon execution
-		pub request: Request<T>,
-		/// Deposit taken for making request (stored in case config changes)
-		pub deposit: BalanceOf<T>,
-		/// All requests expire `T::ExpirationDelay` blocks after they are made
-		pub expires: T::BlockNumber,
-	}
-
-	impl<T: Config> RequestState<T> {
-		fn new(
-			request: Request<T>,
-			deposit: BalanceOf<T>,
-		) -> Result<RequestState<T>, DispatchError> {
-			let expires =
-				frame_system::Pallet::<T>::block_number().saturating_add(T::ExpirationDelay::get());
-			ensure!(
-				request.info.when() < expires,
-				Error::<T>::CannotBeFulfilledBeforeExpiry
-			);
-			Ok(RequestState {
-				request,
-				deposit,
-				expires,
-			})
-		}
-		fn fulfill(&self, caller: &T::AccountId) -> DispatchResult {
-			ensure!(
-				self.request.can_be_fulfilled(),
-				Error::<T>::RequestCannotYetBeFulfilled
-			);
-			let raw_randomness: T::Hash = match self.request.info {
-				RequestType::Local { .. } => return Err(Error::<T>::NotYetImplemented.into()),
-				RequestType::Babe { info, .. } => {
-					Pallet::<T>::get_most_recent_babe_randomness(info)
-				}
-			};
-			let randomness = Pallet::<T>::concat_and_hash(raw_randomness, self.request.salt);
-			T::RandomnessSender::send_randomness(self.request.contract_address.clone(), randomness);
-			// return deposit + fee_excess to contract_address
-			// refund cost_of_execution to caller?
-			T::Currency::unreserve(
-				&self.request.contract_address,
-				self.deposit + self.request.fee,
-			);
-			T::Currency::transfer(
-				&self.request.contract_address,
-				caller,
-				self.request.fee,
-				KeepAlive,
-			)
-			.expect("just unreserved deposit + fee => fee must be transferrable");
-			Ok(())
-		}
-		fn increase_fee(&mut self, caller: &T::AccountId, new_fee: BalanceOf<T>) -> DispatchResult {
-			ensure!(
-				caller == &self.request.contract_address,
-				Error::<T>::OnlyRequesterCanIncreaseFee
-			);
-			let to_reserve = new_fee
-				.checked_sub(&self.request.fee)
-				.ok_or(Error::<T>::NewFeeMustBeGreaterThanOldFee)?;
-			T::Currency::reserve(caller, to_reserve)?;
-			self.request.fee = new_fee;
-			Ok(())
-		}
-		/// Unreserve deposit + fee from contract_address
-		/// Transfer fee to caller
-		fn execute_expiration(&self, caller: &T::AccountId) -> DispatchResult {
-			ensure!(
-				frame_system::Pallet::<T>::block_number() >= self.expires,
-				Error::<T>::RequestHasNotExpired
-			);
-			T::Currency::unreserve(
-				&self.request.contract_address,
-				self.deposit + self.request.fee,
-			);
-			T::Currency::transfer(
-				&self.request.contract_address,
-				caller,
-				self.request.fee,
-				KeepAlive,
-			)
-			.expect("just unreserved deposit + fee => fee must be transferrable");
-			Ok(())
-		}
-	}
-
-	/// Send randomness to a smart contract, triggered by this pallet
-	pub trait SendRandomness<AccountId, R> {
-		fn send_randomness(contract: AccountId, randomness: R);
-	}
-
-	/// Get babe randomness to insert into runtime
-	pub trait GetRelayRandomness<R> {
-		fn get_current_block_randomness() -> (R, Weight);
-		fn get_one_epoch_ago_randomness() -> (R, Weight);
-		fn get_two_epochs_ago_randomness() -> (R, Weight);
-	}
+	use sp_runtime::traits::Saturating;
 
 	/// Request identifier, unique per request for randomness
 	pub type RequestId = u64;
@@ -213,6 +68,8 @@ pub mod pallet {
 		/// Send randomness to smart contract
 		/// TODO: why can't Randomness = T::Hash?
 		type RandomnessSender: SendRandomness<Self::AccountId, [u8; 32]>;
+		/// Convert a weight value into a deductible fee based on the currency type.
+		type WeightToFee: WeightToFeePolynomial<Balance = BalanceOf<Self>>;
 		#[pallet::constant]
 		/// The amount that should be taken as a security deposit when requesting randomness.
 		type Deposit: Get<BalanceOf<Self>>;
@@ -308,10 +165,11 @@ pub mod pallet {
 
 	#[pallet::hooks]
 	impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
+		// Get randomness from runtime and set it locally
 		fn on_initialize(_now: BlockNumberFor<T>) -> Weight {
-			// Get randomness from runtime and set it locally
-			// TODO: optimize by writing a PushRandomness trait which pushes to this pallet when
-			// the item is updated instead...Need to know when the epoch changes from relay proof.
+			// TODO: use `GetEpochNumber` trait associated type to only update some epoch values
+			// upon epoch changes (this will also enable storing epoch index and using it in
+			// `BabeRequestType` when requesting randomness)
 			let (
 				last_current_block_randomness,
 				last_one_epoch_ago_randomness,
@@ -343,14 +201,14 @@ pub mod pallet {
 
 	// Utility functions
 	impl<T: Config> Pallet<T> {
-		fn get_most_recent_babe_randomness(b: BabeRandomness) -> T::Hash {
+		pub(crate) fn get_most_recent_babe_randomness(b: BabeRandomness) -> T::Hash {
 			match b {
 				BabeRandomness::OneEpochAgo => <OneEpochAgoRandomness<T>>::get(),
 				BabeRandomness::TwoEpochsAgo => <TwoEpochsAgoRandomness<T>>::get(),
 				BabeRandomness::CurrentBlock => <CurrentBlockRandomness<T>>::get(),
 			}
 		}
-		fn concat_and_hash(a: T::Hash, b: T::Hash) -> [u8; 32] {
+		pub(crate) fn concat_and_hash(a: T::Hash, b: T::Hash) -> [u8; 32] {
 			let mut s = Vec::new();
 			s.extend_from_slice(a.as_ref());
 			s.extend_from_slice(b.as_ref());
