+++ runtime/moonbase/Cargo.toml
@@ -33,6 +33,7 @@ pallet-maintenance-mode = { path = "../../pallets/maintenance-mode", default-fea
 pallet-migrations = { path = "../../pallets/migrations", default-features = false }
 pallet-proxy-genesis-companion = { path = "../../pallets/proxy-genesis-companion", default-features = false }
 parachain-staking = { path = "../../pallets/parachain-staking", default-features = false }
+pallet-randomness = { path = "../../pallets/randomness", default-features = false }
 xcm-transactor = { path = "../../pallets/xcm-transactor", default-features = false }
 
 # Moonbeam precompiles
@@ -226,6 +227,7 @@ std = [
 	"xcm-transactor/std",
 	"xcm/std",
 	"xtokens-precompiles/std",
+	"pallet-randomness/std",
  ]
 
 # Must be enabled for tracing runtimes only
